refactor(ricer::error)!: remove `error` module
[BREAKING CHANGE]: modules will now define their own custom error types
